requirements.txt
streamlit
faiss-cpu
mistralai
requests
beautifulsoup4
numpy

